# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ETK API",
        "contact": {
            "name": "fbeck",
            "email": "loordy@live.ru"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "First iteration"
        },
        {
            "url": "/api/v2",
            "description": "Second iteration"
        },
        {
            "url": "/api/v1",
            "description": "First iteration"
        },
        {
            "url": "/api/v2",
            "description": "Second iteration"
        }
    ],
    "paths": {
...
        "/position/": {
            "post": {
                "tags": [
                    "Post"
                ],
                "operationId": "/position/",
                "requestBody": {
                    "$ref": "#/components/requestBodies/PositionsStore"
                },
                "responses": {
                    "200": {
                        "description": "Post saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "type": "object",
                                                "$ref": "#/components/schemas/Positions"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Post retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to save this position"
                    },
                    "404": {
                        "description": "Parent position not found"
                    },
                    "424": {
                        "description": "SMTH wrong with dates"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
        "/position/{id}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "operationId": "/position/",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Positions"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Post retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Post not found"
                    },
                    "405": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Post"
                ],
                "operationId": "/position/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/PositionsUpdate"
                },
                "responses": {
                    "200": {
                        "description": "Post saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Positions"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Post retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized to update this position"
                    },
                    "404": {
                        "description": "Parent structure not found"
                    },
                    "424": {
                        "description": "SMTH wrong with dates"
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "operationId": "/position/{id}",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/PositionsDestroy"
                },
                "responses": {
                    "200": {
                        "description": "Post saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Positions"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Post retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
        "/positions/{tin}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "operationId": "/positions/{tin}",
                "parameters": [
                    {
                        "name": "tin",
                        "in": "path",
                        "description": "Status values that needed to be considered for filter",
                        "required": true,
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Post saved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "properties": {
                                            "data": {
                                                "$ref": "#/components/schemas/Positions"
                                            },
                                            "success": {
                                                "type": "string",
                                                "enum": [
                                                    "true",
                                                    "false"
                                                ],
                                                "example": "success"
                                            },
                                            "message": {
                                                "type": "string",
                                                "example": "Post retrieved successfully"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "etk_auth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Positions": {
                "title": "Структура позиций",
                "description": "Структура позиций",
                "required": [
                    "id_position"
                ],
                "properties": {
                    "id_position": {
                        "title": "Уникальный идентификатор должности",
                        "description": "Уникальный идентификатор должности",
                        "format": "int64"
                    },
                    "id_structure": {
                        "title": "Уникальный идентификатор структуры",
                        "description": "Уникальный идентификатор структуры",
                        "format": "int64"
                    },
                    "name_position": {
                        "title": "Наименование должности",
                        "description": "Наименование должности",
                        "format": "string"
                    },
                    "code_prof_position": {
                        "title": "Код профессии (КОДП)",
                        "description": "Код профессии (КОДП)",
                        "format": "string"
                    },
                    "salary_position": {
                        "title": "Оклад должностной",
                        "description": "Оклад должностной",
                        "format": "string"
                    },
                    "requirements_position": {
                        "title": "Требования к должности",
                        "description": "Требования к должности",
                        "format": "string"
                    },
                    "duties_position": {
                        "title": "Обязанности должности",
                        "description": "Обязанности должности",
                        "format": "string"
                    },
                    "conditions_position": {
                        "title": "Условия должности",
                        "description": "Условия должности",
                        "format": "string"
                    },
                    "id_vacancy_position": {
                        "title": "Идентификатор вакансии в хранилище",
                        "description": "Идентификатор вакансии в хранилище",
                        "format": "int64"
                    },
                    "form_position": {
                        "title": "Вид занятости РМ",
                        "description": "Вид занятости РМ",
                        "format": "int2"
                    },
                    "status_open_position": {
                        "title": "Позиция не занята",
                        "description": "Позиция не занята",
                        "format": "boolean"
                    },
                    "date_end_position": {
                        "title": "Дата конца",
                        "description": "Позиция не занята",
                        "format": "date"
                    },
                    "flag_bonus_position": {
                        "title": "Признак надбавка",
                        "description": "Признак надбавка",
                        "format": "boolean"
                    },
                    "type_emp_position": {
                        "title": "Тип занятости",
                        "description": "Тип занятости(описание в константах)",
                        "format": "integer"
                    },
                    "code_prof_type_position": {
                        "title": "Дополнительный идентификатор по кодп",
                        "description": "Дополнительный идентификатор по кодп(type в таблице кодп)",
                        "format": "string"
                    },
                    "rate_position": {
                        "title": "Ставка должности",
                        "description": "Ставка должности",
                        "format": "numeric"
                    },
                    "date_start_position": {
                        "title": "Дата начала",
                        "description": "Позиция не занята",
                        "format": "date"
                    }
                },
                "type": "object",
                "example": {
                    "status_open_position": true,
                    "id_structure_position": 5,
                    "name_position": "POS4_3",
                    "date_end_position": null,
                    "date_start_position": "2018-08-08 00:00:00.000000",
                    "requirements_position": null,
                    "form_position": null,
                    "code_prof_position": "12    ",
                    "salary_position": "1030000.00",
                    "id_position": 6337,
                    "id_vacancy_position": 3961
                }
            },
...
        },
        "requestBodies": {
...
            "PositionsStore": {
                "description": "Department object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "items": {
                                "title": "Структура организации",
                                "description": "Структура организации",
                                "required": [
                                    "id_structure_position",
                                    "name_position",
                                    "date_start_position",
                                    "code_prof_position",
                                    "salary_position"
                                ],
                                "example": {
                                    "id_structure_position": 5,
                                    "name_position": "POS4_3",
                                    "date_end_position": "2018-12-08 00:00:00",
                                    "date_start_position": "2018-08-08 00:00:00.000000",
                                    "status_open_position": true,
                                    "requirements_position": "",
                                    "form_position": 1,
                                    "code_prof_position": "12",
                                    "salary_position": "1030000.00"
                                }
                            }
                        }
                    }
                }
            },
            "PositionsUpdate": {
                "description": "Department object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "items": {
                                "title": "Структура организации",
                                "description": "Структура организации",
                                "required": [],
                                "example": {
                                    "id_structure_position": 5,
                                    "name_position": "POS4_3",
                                    "date_end_position": "2018-12-08 00:00:00",
                                    "date_start_position": "2018-08-08 00:00:00",
                                    "status_open_position": true,
                                    "requirements_position": "",
                                    "form_position": 1,
                                    "code_prof_position": "12"
                                }
                            }
                        }
                    }
                }
            },
            "PositionsDestroy": {
                "description": "Department object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "items": {
                                "title": "Структура организации",
                                "description": "Структура организации",
                                "required": [],
                                "example": []
                            }
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "etk_auth": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "refreshUrl": "/oauth/token",
                        "scopes": []
                    }
                }
            }
        }
    },
...
}